requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.5
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.13.4
cryptography==46.0.5
daphne==4.2.1
distro==1.9.0
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-safedelete==1.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
groq==1.0.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
whitenoise==6.11.0
zope.interface==8.2
